refactor: 위임만 하는 클래스 삭제
